pyproject.toml
[project]
name = "indextts-fastapi"
version = "1.0.0"
description = "FastAPI REST API wrapper for IndexTTS2 with OpenAI-compatible endpoints"
authors = [{ name = "IndexTTS FastAPI Contributors" }]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "pydub>=0.25.1",
    "pyyaml>=6.0.0",
    "huggingface_hub[cli,hf_xet]>=0.30.0,<1.0",  # For model downloading, pinned to match transformers==4.52.1 requirement
    # Note: indextts is installed from cloned repo, not from PyPI
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
indextts-api = "indextts_fastapi.api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["indextts_fastapi"]

